Fix invalid grid padding on android
