+++ packages/react-native-sortables/src/components/SortableGrid.tsx
@@ -234,10 +234,7 @@ function SortableGridComponent<I>({
           }
         : {
             flexBasis: mainGroupSize.value ? undefined : `${100 / groups}%`,
-            paddingHorizontal: columnGap.value / 2,
-            width: mainGroupSize.value
-              ? mainGroupSize.value + columnGap.value
-              : undefined
+            paddingHorizontal: columnGap.value / 2
           }
       : { height: rowHeight }
   );
