+++ packages/react-native-sortables/src/types/providers/shared.ts
@@ -1,11 +1,12 @@
 import type { ReactNode } from 'react';
-import type { ScrollView, View } from 'react-native';
+import type { ScrollView, View, ViewStyle } from 'react-native';
 import type {
   GestureTouchEvent,
   GestureType
 } from 'react-native-gesture-handler';
 import type {
   AnimatedRef,
+  AnimatedStyle,
   MeasuredDimensions,
   SharedValue
 } from 'react-native-reanimated';
@@ -79,6 +80,7 @@ export type CommonValuesContextType =
       itemWidths: SharedValue<ItemSizes>;
       itemHeights: SharedValue<ItemSizes>;
       activeItemDimensions: SharedValue<Dimensions | null>;
+      controlledItemDimensionsStyle: AnimatedStyle<ViewStyle>;
 
       // DRAG STATE
       prevActiveItemKey: SharedValue<null | string>;
